+++ packages/ts-compiler/test/test-files.spec.ts
@@ -1,6 +1,6 @@
 import { buildComponentInWorkspace, getDefaultComponent } from './build-default-component';
 import Helper from 'bit-bin/dist/e2e-helper/e2e-helper';
-import { GenericObject } from '../src/compiler';
+import { GenericObject } from '@bit/bit.envs.common.compiler-types';
 import { buildOne } from 'bit-bin';
 import { expect } from 'chai';
 import compiler from './spy-compiler';
